Fix type inference

+++ src/gt4py/next/iterator/pretty_parser.py
@@ -101,7 +101,7 @@ class ToIrTransformer(lark_visitors.Transformer):
         return im.literal(int(value))
 
     def FLOAT_LITERAL(self, value: lark_lexer.Token) -> ir.Literal:
-        return ir.Literal(value=value.value, type="float")
+        return ir.Literal(value=value.value, type="float64")
 
     def OFFSET_LITERAL(self, value: lark_lexer.Token) -> ir.OffsetLiteral:
         v: Union[int, str] = value.value[:-1]
